Fetch MP identifiers
ScraperWiki division pages don’t include their unique ID for MPs, but
only a URL constructed from their name and constituency. So look that
up, and store it in another table we can cross match with.

This isn’t really the ideal structure/approach to this, but it’s Good
Enough For Now™.